Use moto for region test
