+++ tests/test_aws.py
@@ -1,11 +1,12 @@
-from geopandas import GeoDataFrame
-from shapely import box
-
 from dep_tools.aws import get_s3_bucket_region
 
+TEST_BUCKET_NAME = "dep-test-bucket"
 
-def test_get_s3_bucket_region():
-    region = get_s3_bucket_region("dep-public-staging")
+
+def test_get_s3_bucket_region(s3):
+    s3.create_bucket(Bucket=TEST_BUCKET_NAME, CreateBucketConfiguration={
+        'LocationConstraint': 'us-west-2'})
+    region = get_s3_bucket_region(TEST_BUCKET_NAME)
     assert region == "us-west-2"
