+++ tests/test_namers.py
@@ -1,8 +1,3 @@
-import os
-
-import boto3
-import moto
-import pytest
 
 from dep_tools.namers import DailyItemPath, DepItemPath
 
@@ -18,25 +13,6 @@ nonPaddedItemPath = DepItemPath(
 )
 item_id = "001,002"
 asset_name = "mean"
-
-
-@pytest.fixture(scope="function")
-def aws_credentials():
-    """Mocked AWS Credentials for moto."""
-    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
-    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
-    os.environ["AWS_SECURITY_TOKEN"] = "testing"
-    os.environ["AWS_SESSION_TOKEN"] = "testing"
-    os.environ["AWS_DEFAULT_REGION"] = "us-east-1"
-
-
-@pytest.fixture(scope="function")
-def s3():
-    """
-    Return a mocked S3 client
-    """
-    with moto.mock_aws():
-        yield boto3.client("s3", region_name="us-east-1")
 
 
 def test_path():
